requirements.txt
gradio
ultralytics
opencv-python
numpy
reportlab
openpyxl
Pillow
